+++ R/table_to_parquet.R
@@ -16,24 +16,25 @@
 #'
 #' }
 #'
-#' To avoid overcharging R's RAM, the conversion can be done by chunk. One of arguments `chunk_memory_size` or `chunk_size` must then be used.
+#' To avoid overcharging R's RAM, the conversion can be done by chunk. One of arguments `max_memory` or `max_rows` must then be used.
 #' This is very useful for huge tables and for computers with little RAM because the conversion is then done
 #' with less memory consumption. For more information, see [here](https://ddotta.github.io/parquetize/articles/aa-conversions.html).
 #'
 #' @param path_to_table String that indicates the path to the input file (don't forget the extension).
 #' @param path_to_parquet String that indicates the path to the directory where the parquet files will be stored.
 #' @param columns Character vector of columns to select from the input file (by default, all columns are selected).
-#' @param by_chunk DEPRECATED use chunk_memory_size or chunk_memory instead
-#' @param chunk_memory_size Memory size (in Mb) in which data of one parquet file should roughly fit. For very small size, data could be a bit larger than given memory.
-#' @param chunk_size Number of lines that defines the size of the chunk.
-#' This argument must be filled in if `by_chunk` is TRUE (otherwise ignored).
-#' This argument can not be filled in if chunk_memory_size is used.
+#' @param max_memory Memory size (in Mb) in which data of one parquet file should roughly fit.
+#' @param max_rows Number of lines that defines the size of the chunk.
+#' This argument can not be filled in if max_memory is used.
+#' @param by_chunk DEPRECATED use max_memory or max_rows instead
+#' @param chunk_size DEPRECATED use max_rows
+#' @param chunk_memory_size DEPRECATED use max_memory
 #' @param skip By default 0. This argument must be filled in if `by_chunk` is TRUE. Number of lines to ignore when converting.
 #' @param partition String ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file.
 #' If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.
-#' Be careful, this argument can not be "yes" if `chunk_memory_size` or `chunk_size` argument are not NULL.
+#' Be careful, this argument can not be "yes" if `max_memory` or `max_rows` argument are not NULL.
 #' @param encoding String that indicates the character encoding for the input file.
-#' @param chunk_memory_sample_lines Number of lines to read to evaluate chunk_memory_size. Default to 10 000.
+#' @param chunk_memory_sample_lines Number of lines to read to evaluate max_memory. Default to 10 000.
 #' @param ... Additional format-specific arguments,  see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}
 #'  and \href{https://arrow.apache.org/docs/r/reference/write_dataset.html}{arrow::write_dataset()} for more informations.
 #'
@@ -68,16 +69,16 @@
 #'   path_to_parquet = tempdir()
 #' )
 #'
-#' # Reading SPSS file by chunk (using `chunk_size` argument)
+#' # Reading SPSS file by chunk (using `max_rows` argument)
 #' # and conversion to multiple parquet files :
 #'
 #' table_to_parquet(
 #'   path_to_table = system.file("examples","iris.sav", package = "haven"),
 #'   path_to_parquet = tempdir(),
-#'   chunk_size = 50,
+#'   max_rows = 50,
 #' )
 #'
-#' # Reading SPSS file by chunk (using `chunk_memory_size` argument)
+#' # Reading SPSS file by chunk (using `max_memory` argument)
 #' # and conversion to multiple parquet files of 5 Kb when loaded (5 Mb / 1024)
 #' # (in real files, you should use bigger value that fit in memory like 3000
 #' # or 4000) :
@@ -85,7 +86,7 @@
 #' table_to_parquet(
 #'   path_to_table = system.file("examples","iris.sav", package = "haven"),
 #'   path_to_parquet = tempdir(),
-#'   chunk_memory_size = 5 / 1024,
+#'   max_memory = 5 / 1024,
 #' )
 #'
 #' # Reading SAS file by chunk of 50 lines with encoding
@@ -94,7 +95,7 @@
 #' table_to_parquet(
 #'   path_to_table = system.file("examples","iris.sas7bdat", package = "haven"),
 #'   path_to_parquet = tempdir(),
-#'   chunk_size = 50,
+#'   max_rows = 50,
 #'   encoding = "utf-8"
 #' )
 #'
@@ -123,7 +124,7 @@
 #' table_to_parquet(
 #'   path_to_table = system.file("examples","iris.sas7bdat", package = "haven"),
 #'   path_to_parquet = tempdir(),
-#'   chunk_size = 50,
+#'   max_rows = 50,
 #'   compression = "zstd",
 #'   compression_level = 10
 #' )
@@ -132,8 +133,10 @@
 table_to_parquet <- function(
     path_to_table,
     path_to_parquet,
-    chunk_memory_size,
-    chunk_size,
+    max_memory,
+    max_rows,
+    chunk_size = lifecycle::deprecated(),
+    chunk_memory_size = lifecycle::deprecated(),
     columns = "all",
     by_chunk = lifecycle::deprecated(),
     skip = 0,
@@ -146,8 +149,26 @@ table_to_parquet <- function(
     lifecycle::deprecate_warn(
       when = "0.5.5",
       what = "table_to_parquet(by_chunk)",
-      details = "This argument is no longer needed, table_to_parquet will chunk if one of chunk_memory_size or chunk_size is setted"
+      details = "This argument is no longer needed, table_to_parquet will chunk if one of max_memory or max_rows is setted"
     )
+  }
+
+  if (!missing(chunk_size)) {
+    lifecycle::deprecate_warn(
+      when = "0.5.5",
+      what = "table_to_parquet(chunk_size)",
+      details = "This argument is deprecated, use max_rows."
+    )
+    max_rows <- chunk_size
+  }
+
+  if (!missing(chunk_memory_size)) {
+    lifecycle::deprecate_warn(
+      when = "0.5.5",
+      what = "table_to_parquet(chunk_memory_size)",
+      details = "This argument is deprecated, use max_memory."
+    )
+    max_memory <- chunk_memory_size
   }
 
   # Check if path_to_table is missing
@@ -169,7 +190,7 @@ table_to_parquet <- function(
     class = "parquetize_bad_type")
   }
 
-  by_chunk <- !(missing(chunk_size) & missing(chunk_memory_size))
+  by_chunk <- !(missing(max_rows) & missing(max_memory))
 
   # Check if skip argument is correctly filled in by_chunk argument is TRUE
   if (by_chunk==TRUE & skip<0) {
@@ -179,21 +200,21 @@ table_to_parquet <- function(
 
   # If by_chunk argument is TRUE and partition argument is equal to "yes" it fails
   if (by_chunk==TRUE & partition == "yes") {
-    cli_abort("Be careful, when by_chunk is TRUE partition and partitioning can not be used", class = "parquetize_bad_argument")
+    cli_abort("Be careful, when max_rows or max_memory are used, partition and partitioning can not be used", class = "parquetize_bad_argument")
   }
 
-  # chunk_size and chunk_memory_size can not be used together so fails
-  if (!missing(chunk_size) & !missing(chunk_memory_size)) {
-    cli_abort("Be careful, chunk_size and chunk_memory_size can not be used together", class = "parquetize_bad_argument")
+  # max_rows and max_memory can not be used together so fails
+  if (!missing(max_rows) & !missing(max_memory)) {
+    cli_abort("Be careful, max_rows and max_memory can not be used together", class = "parquetize_bad_argument")
   }
 
   if (by_chunk) {
     read_method <- get_read_function_for_file(path_to_table)
 
-    if (missing(chunk_size)) {
+    if (missing(max_rows)) {
       data <- read_method(path_to_table, n_max = chunk_memory_sample_lines)
-      chunk_size <- get_lines_for_memory(data,
-                                         chunk_memory_size = chunk_memory_size)
+      max_rows <- get_lines_for_memory(data,
+                                         max_memory = max_memory)
     }
 
     parquetname <- paste0(gsub("\\..*","",basename(path_to_table)))
@@ -202,8 +223,9 @@ table_to_parquet <- function(
     while (TRUE) {
       tbl <- read_method(path_to_table,
                          skip = skip,
-                         n_max = chunk_size)
-
+                         n_max = max_rows,
+                         encoding = encoding,
+                         col_select = if (identical(columns,"all")) everything() else all_of(columns))
       if (nrow(tbl) != 0) {
         parquetizename <- paste0(parquetname,sprintf("%d",skip+1),"-",sprintf("%d",skip+nrow(tbl)),".parquet")
         write_parquet(tbl,
@@ -213,9 +235,9 @@ table_to_parquet <- function(
         )
         cli_alert_success("\nThe {get_file_format(path_to_table)} file is available in parquet format under {path_to_parquet}/{parquetizename}")
       }
-      skip <- skip + chunk_size
+      skip <- skip + max_rows
 
-      if (nrow(tbl) < chunk_size) { return(invisible(TRUE)) }
+      if (nrow(tbl) < max_rows) { return(invisible(TRUE)) }
     }
   }
   read_method <- get_read_function_for_file(path_to_table)